+++ src/components/GuestReviews/GuestReviews.test.js
@@ -39,4 +39,11 @@ describe('GuestReviews', () => {
       )
     ).toMatchSnapshot();
   });
+
+  // describe('if there are still something to load next time', () => {
+  //
+  // });
+  // describe('if there is nothing left to load next time', () => {
+  //
+  // });
 });